Selection Toolbar
Add a selection toolbar NSView.
Add a selection action object to define what toolbar buttons do
Add a button object that represent the above action to be contained in the selection toolbar
Add the toolbar to the selection. The toolbar appears in the upper right corner of the selection, unless the selection isn't big enough. Then it goes to the right or left outside edge depending on whether the selection end is close to the right hand edge of the view.
Add methods to MLNSampleViewDelegate for obtaining the contents of the toolbar (and a future sample view menu)