Don’t let the overview bar’s visible region ends to overwrap
Dragging a handle past the other handles caused the visible region to get screwed up. Special care needed to be taken when the requested visible range was a lot smaller than the current range, as the sample view would not have had time to resize to the correct size, causing the scrollPoint message to put the scrollview in an odd position. To fix that we force the view to be laid out before attempting to scroll